+++ assets/FamilyDirectoryPdfGeneratorLambda/src/main/java/org/familydirectory/assets/lambda/function/stream/FamilyDirectoryPdfGeneratorLambda.java
@@ -4,6 +4,10 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.DynamodbEvent;
 import com.amazonaws.services.lambda.runtime.logging.LogLevel;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
 import org.familydirectory.assets.lambda.function.stream.helper.PdfHelper;
 import org.familydirectory.assets.lambda.function.utility.LambdaUtils;
 import org.jetbrains.annotations.NotNull;
@@ -11,6 +15,7 @@ import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import static java.lang.System.getenv;
+import static java.util.Objects.requireNonNull;
 
 public
 class FamilyDirectoryPdfGeneratorLambda implements RequestHandler<DynamodbEvent, Void> {
@@ -22,18 +27,17 @@ class FamilyDirectoryPdfGeneratorLambda implements RequestHandler<DynamodbEvent,
             final String pdfKey;
             final RequestBody pdfRequestBody;
             try (final PdfHelper pdfHelper = new PdfHelper(context.getLogger())) {
-                pdfKey = pdfHelper.getPdfS3Key();
-                pdfRequestBody = pdfHelper.getPdf();
-            } catch (final Throwable e) {
-                context.getLogger()
-                       .log("FATAL: Issue Exists in PdfHelper/PDPageHelper", LogLevel.FATAL);
+                final String rootMemberSurname = pdfHelper.getRootMemberSurname();
+                pdfKey = pdfHelper.getPdfS3Key(rootMemberSurname);
+                pdfRequestBody = zipPdfBundle(pdfHelper, rootMemberSurname);
+            } catch (final IOException e) {
                 throw new RuntimeException(e);
             }
 
             final PutObjectRequest pdfPutRequest = PutObjectRequest.builder()
                                                                    .bucket(getenv(LambdaUtils.EnvVar.S3_PDF_BUCKET_NAME.name()))
                                                                    .key(pdfKey)
-                                                                   .contentType("application/pdf")
+                                                                   .contentType("application/zip")
                                                                    .build();
             try (final S3Client s3Client = S3Client.create()) {
                 s3Client.putObject(pdfPutRequest, pdfRequestBody);
@@ -44,6 +48,25 @@ class FamilyDirectoryPdfGeneratorLambda implements RequestHandler<DynamodbEvent,
         } catch (final Throwable e) {
             LambdaUtils.logTrace(context.getLogger(), e, LogLevel.FATAL);
             throw e;
+        }
+    }
+
+    private static @NotNull
+    RequestBody zipPdfBundle (final @NotNull PdfHelper pdfHelper, final @NotNull String rootMemberSurname) throws IOException {
+        final PdfHelper.PdfBundle pdfBundle = requireNonNull(pdfHelper).getPdfBundle();
+        try (final ByteArrayOutputStream bos = new ByteArrayOutputStream(); final ZipOutputStream zos = new ZipOutputStream(bos)) {
+            final ZipEntry familyDirectoryZip = new ZipEntry("%sFamilyDirectory.pdf".formatted(requireNonNull(rootMemberSurname)));
+            zos.putNextEntry(familyDirectoryZip);
+            zos.write(pdfBundle.familyDirectoryPdf());
+            zos.closeEntry();
+
+            final ZipEntry birthdayZip = new ZipEntry("%sFamilyBirthdays.pdf".formatted(rootMemberSurname));
+            zos.putNextEntry(birthdayZip);
+            zos.write(pdfBundle.birthdayPdf());
+            zos.closeEntry();
+
+            zos.close();
+            return RequestBody.fromBytes(bos.toByteArray());
         }
     }
 }
